coop: separate agent protocol responsibilities
