+++ pkg/cmd/coop/coop_agent.go
@@ -1,17 +1,9 @@
 package coopcmd
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"os"
-	"strings"
-
 	"github.com/spf13/cobra"
 
 	"github.com/stripe/stripe-cli/pkg/coop"
-	"github.com/stripe/stripe-cli/pkg/coop/followups"
-	"github.com/stripe/stripe-cli/pkg/coop/helpers"
 	"github.com/stripe/stripe-cli/pkg/coop/workflow"
 )
 
@@ -200,287 +192,4 @@ func newCoopAgentAwaitReviewCmd() *coopAgentActionCmd {
 	c.addSessionStepFlags()
 	configureAgentCommand(c.cmd)
 	return c
-}
-
-func newCoopAgentNextActionCmd() *coopAgentActionCmd {
-	c := &coopAgentActionCmd{}
-	c.cmd = &cobra.Command{
-		Use:   "next-action",
-		Short: "Wait for or record the developer's next action",
-		Args:  agentNoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if strings.TrimSpace(c.session) == "" {
-				return outputCoopError(
-					"--session flag is required",
-					"Retry next-action with the intended session.",
-					coop.NextActionTemplate(),
-				)
-			}
-			return runCoopNextAction(c.session, c.completed)
-		},
-	}
-	c.cmd.Flags().StringVar(&c.session, "session", "", "Session ID")
-	c.cmd.Flags().StringVar(&c.completed, "completed", "", "Mark a next action as completed")
-	configureAgentCommand(c.cmd)
-	return c
-}
-
-func newCoopAgentStartFollowupCmd() *coopAgentActionCmd {
-	c := &coopAgentActionCmd{ensureSkill: ensureRepoStripeBestPracticesSkill}
-	c.cmd = &cobra.Command{
-		Use:   "start-followup",
-		Short: "Start an internal guided follow-up session",
-		Args:  agentNoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if strings.TrimSpace(c.session) == "" || strings.TrimSpace(c.action) == "" {
-				return outputCoopError(
-					"--session and --action flags are required",
-					"Provide the parent session and an offered follow-up action.",
-					coop.StartFollowupTemplate(""),
-				)
-			}
-			return runCoopStartFollowup(cmd, c.session, c.action, c.target, c.ensureSkill)
-		},
-	}
-	c.cmd.Flags().StringVar(&c.session, "session", "", "Parent session ID")
-	c.cmd.Flags().StringVar(&c.action, "action", "", "Follow-up action ID")
-	c.cmd.Flags().StringVar(&c.target, "target", "", "Detected deployment target")
-	configureAgentCommand(c.cmd)
-	return c
-}
-
-func (c *coopAgentActionCmd) addSessionStepFlags() {
-	c.cmd.Flags().StringVar(&c.session, "session", "", "Session ID")
-	c.cmd.Flags().IntVar(&c.step, "step", 0, "1-based node number")
-}
-
-func (c *coopAgentActionCmd) validateSessionStep(action string) error {
-	if strings.TrimSpace(c.session) != "" && c.step > 0 {
-		return nil
-	}
-	return outputCoopError(
-		"--session and a positive --step are required",
-		"Provide the Co-op session ID and 1-based node number.",
-		coop.SessionStepTemplate(action),
-	)
-}
-
-func configureAgentCommand(cmd *cobra.Command) {
-	cmd.SilenceErrors = true
-	cmd.SilenceUsage = true
-	cmd.SetFlagErrorFunc(func(cmd *cobra.Command, err error) error {
-		return outputCoopError(
-			err.Error(),
-			"Correct the command flags and retry.",
-			coop.Continue(coop.StatusCommand("")),
-		)
-	})
-}
-
-func agentNoArgs(cmd *cobra.Command, args []string) error {
-	if len(args) == 0 {
-		return nil
-	}
-	return outputCoopError(
-		fmt.Sprintf("%s does not accept positional arguments", cmd.CommandPath()),
-		"Remove the unexpected positional arguments and retry.",
-		coop.Continue(coop.StatusCommand("")),
-	)
-}
-
-func newWorkflowService() (*workflow.Service, error) {
-	store, err := coop.NewStore(coopConfigFolder())
-	if err != nil {
-		return nil, fmt.Errorf("creating store: %w", err)
-	}
-	return workflow.NewService(store), nil
-}
-
-func parseReportedOutputs(values []string) (coop.NodeOutputs, error) {
-	if len(values) == 0 {
-		return nil, nil
-	}
-	outputs := coop.NodeOutputs{}
-	for _, value := range values {
-		selector, rawValue, ok := strings.Cut(value, "=")
-		if !ok {
-			return nil, fmt.Errorf("--output must be in field=value or source:field=value format: %q", value)
-		}
-		selector = strings.TrimSpace(selector)
-		if selector == "" {
-			return nil, fmt.Errorf("--output field cannot be empty: %q", value)
-		}
-		if rawValue == "" {
-			return nil, fmt.Errorf("--output value cannot be empty: %q", value)
-		}
-
-		source := coop.DefaultOutputSource
-		field := selector
-		if parsedSource, parsedField, hasSource := strings.Cut(selector, ":"); hasSource {
-			source = strings.TrimSpace(parsedSource)
-			field = strings.TrimSpace(parsedField)
-			if source == "" || field == "" {
-				return nil, fmt.Errorf("--output source and field cannot be empty: %q", value)
-			}
-		}
-
-		raw := json.RawMessage(rawValue)
-		if !json.Valid(raw) {
-			encoded, err := json.Marshal(rawValue)
-			if err != nil {
-				return nil, fmt.Errorf("encoding --output %q: %w", selector, err)
-			}
-			raw = encoded
-		}
-		if outputs[source] == nil {
-			outputs[source] = map[string]json.RawMessage{}
-		}
-		outputs[source][field] = append(json.RawMessage(nil), raw...)
-	}
-	return outputs, nil
-}
-
-func runCoopNextAction(sessionID, completed string) error {
-	store, err := coop.NewStore(coopConfigFolder())
-	if err != nil {
-		return outputAgentError(fmt.Errorf("creating store: %w", err))
-	}
-	return runCoopNextActionWithStore(store, sessionID, completed)
-}
-
-func runCoopNextActionWithStore(store helpers.Store, sessionID, completed string) error {
-	resp, err := helpers.Run(store, helpers.Input{SessionID: sessionID, Completed: completed})
-	if errors.Is(err, helpers.ErrNoSession) {
-		return outputCoopError(
-			"No session found.",
-			"Start a Co-op session before requesting a next action.",
-			coop.RunTemplate(),
-		)
-	}
-	if err != nil {
-		return outputCoopError(
-			err.Error(),
-			"Retry the next-action wait.",
-			coop.Continue(coop.NextActionCommand(sessionID, "")),
-		)
-	}
-	return outputJSON(resp)
-}
-
-func runCoopStartFollowup(cmd *cobra.Command, parentSessionID, actionID, target string, ensureSkill func() error) error {
-	store, err := coop.NewStore(coopConfigFolder())
-	if err != nil {
-		return outputAgentError(fmt.Errorf("creating store: %w", err))
-	}
-
-	parent, err := store.Read(parentSessionID)
-	if err != nil {
-		return outputCoopError(
-			fmt.Sprintf("Parent session %q not found.", parentSessionID),
-			"Inspect active and completed Co-op sessions.",
-			coop.Continue(coop.StatusCommand("")),
-		)
-	}
-
-	action, err := followups.GuidedActionByID(actionID, target)
-	if err != nil {
-		return outputCoopError(
-			err.Error(),
-			"Use an action offered by the parent session.",
-			coop.StartFollowupTemplate(parentSessionID),
-		)
-	}
-	if err := validateFollowupParent(parent, action.ID); err != nil {
-		return outputCoopError(
-			err.Error(),
-			"Return to the parent session's next-action selection.",
-			coop.Continue(coop.NextActionCommand(parent.ID, "")),
-		)
-	}
-	if ensureSkill != nil {
-		if err := ensureSkill(); err != nil {
-			warnRepoStripeBestPracticesSkill(cmd, err)
-		}
-	}
-
-	settings := make(map[string]string, len(parent.Settings)+1)
-	for key, value := range parent.Settings {
-		settings[key] = value
-	}
-	if target != "" {
-		settings["deploy_target"] = target
-	}
-
-	sessionID := "coop_" + generateShortID()
-	session := coop.NewSessionFromGuidedAction(action, sessionID, coop.GuidedActionSessionOptions{
-		ParentSessionID: parent.ID,
-		ParentStepID:    action.ID,
-		Settings:        settings,
-		UsedSandbox:     parent.UsedSandbox || coopSandboxClaimURL() != "",
-	})
-	if err := store.Write(session); err != nil {
-		return outputAgentError(fmt.Errorf("writing guided follow-up session: %w", err))
-	}
-
-	return outputJSON(newCoopAgentGuidedActionResponse(action, session))
-}
-
-func validateFollowupParent(parent *coop.Session, actionID string) error {
-	if parent.Status != coop.SessionCompleted {
-		return fmt.Errorf("parent session %q is not completed", parent.ID)
-	}
-	if parent.NextSteps == nil {
-		return fmt.Errorf("parent session %q has no next-step suggestions", parent.ID)
-	}
-	for _, completed := range parent.NextSteps.Completed {
-		if completed == actionID {
-			return fmt.Errorf("follow-up action %q is already completed for parent session %q", actionID, parent.ID)
-		}
-	}
-	for _, suggestion := range parent.NextSteps.Suggestions {
-		if suggestion.ID == actionID {
-			return nil
-		}
-	}
-	return fmt.Errorf("follow-up action %q is not available for parent session %q", actionID, parent.ID)
-}
-
-// outputAgentError renders err as a structured agent JSON response. Used for
-// failures that happen before a workflow CommandResponse exists (e.g.
-// newWorkflowService / store creation), so agent commands never emit a bare
-// plain-text error on that path.
-func outputAgentError(err error) error {
-	return outputAgentResponse(coop.CommandResponse{}, err)
-}
-
-func outputAgentResponse(resp coop.CommandResponse, err error) error {
-	if err != nil {
-		resp = protocolFailure(err.Error())
-	}
-	if validationErr := resp.Validate(); validationErr != nil {
-		resp = protocolFailure("invalid Co-op protocol response: " + validationErr.Error())
-	}
-	if !resp.OK {
-		if resp.Recovery == nil {
-			resp.Recovery = defaultAgentRecovery()
-		}
-		if outErr := outputJSONTo(os.Stderr, resp); outErr != nil {
-			return outErr
-		}
-		return RenderedError{}
-	}
-	return outputJSON(resp)
-}
-
-func protocolFailure(message string) coop.CommandResponse {
-	return coop.CommandResponse{
-		OK:       false,
-		Error:    message,
-		Recovery: defaultAgentRecovery(),
-	}
-}
-
-func defaultAgentRecovery() *coop.Recovery {
-	return coop.Continue(coop.StatusCommand("")).
-		Recovery("Inspect the current Co-op session before retrying.")
 }
